+++ test/spec/vast_player.directive.ut.js
@@ -616,6 +616,20 @@ define(['videos/vast'], function(vastModule) {
                 scope.clickThrough();
                 expect(_player.player.play).toHaveBeenCalled();
             });
+
+            it('should do nothing if click through url is null.com', function() {
+                vastObject.clickThrough[0] = 'http://null.com';
+
+                scope.clickThrough();
+                expect($window.open).not.toHaveBeenCalled();
+            });
+
+            it('should do nothing if click through url is not defined', function() {
+                vastObject.clickThrough = [];
+
+                scope.clickThrough();
+                expect($window.open).not.toHaveBeenCalled();
+            });
         });
 
         describe('$watch', function() {